+++ core/src/services/dropbox/error.rs
@@ -17,19 +17,26 @@
 
 use http::Response;
 use http::StatusCode;
+use http::Uri;
+use serde::Deserialize;
 
-use super::response::DropboxErrorResponse;
 use crate::raw::*;
 use crate::Error;
 use crate::ErrorKind;
 use crate::Result;
+
+#[derive(Default, Debug, Deserialize)]
+#[serde(default)]
+pub struct DropboxErrorResponse {
+    error_summary: String,
+}
 
 /// Parse error response into Error.
 pub async fn parse_error(resp: Response<IncomingAsyncBody>) -> Result<Error> {
     let (parts, body) = resp.into_parts();
     let bs = body.bytes().await?;
 
-    let (kind, retryable) = match parts.status {
+    let (mut kind, mut retryable) = match parts.status {
         StatusCode::NOT_FOUND => (ErrorKind::NotFound, false),
         StatusCode::FORBIDDEN => (ErrorKind::PermissionDenied, false),
         StatusCode::INTERNAL_SERVER_ERROR
@@ -39,53 +46,39 @@ pub async fn parse_error(resp: Response<IncomingAsyncBody>) -> Result<Error> {
         _ => (ErrorKind::Unexpected, false),
     };
 
-    let dropbox_error =
-        serde_json::from_slice::<DropboxErrorResponse>(&bs).map_err(new_json_deserialize_error);
-    match dropbox_error {
-        Ok(dropbox_error) => {
-            // We cannot get the error type from the response header when the status code is 409.
-            // Because Dropbox API v2 will put error summary in the response body,
-            // we need to parse it to get the correct error type and then error kind.
-            // See https://www.dropbox.com/developers/documentation/http/documentation#error-handling
-            let error_summary = dropbox_error.error_summary.as_str();
+    let (message, dropbox_err) = serde_json::from_slice::<DropboxErrorResponse>(&bs)
+        .map(|dropbox_err| (format!("{dropbox_err:?}"), Some(dropbox_err)))
+        .unwrap_or_else(|_| (String::from_utf8_lossy(&bs).into_owned(), None));
 
-            let mut err = Error::new(
-                match parts.status {
-                    // 409 Conflict means that Endpoint-specific error.
-                    // Look to the JSON response body for the specifics of the error.
-                    StatusCode::CONFLICT => {
-                        if error_summary.contains("path/not_found")
-                            || error_summary.contains("path_lookup/not_found")
-                        {
-                            ErrorKind::NotFound
-                        } else if error_summary.contains("path/conflict") {
-                            ErrorKind::AlreadyExists
-                        } else {
-                            ErrorKind::Unexpected
-                        }
-                    }
-                    // Otherwise, we can get the error type from the response status code.
-                    _ => kind,
-                },
-                error_summary,
-            )
-            .with_context("response", format!("{parts:?}"));
+    if let Some(dropbox_err) = dropbox_err {
+        (kind, retryable) =
+            parse_dropbox_error_summary(&dropbox_err.error_summary).unwrap_or((kind, retryable));
+    }
 
-            if retryable {
-                err = err.set_temporary();
-            }
+    let mut err = Error::new(kind, &message).with_context("response", format!("{parts:?}"));
 
-            Ok(err)
-        }
-        Err(_err) => {
-            let mut err = Error::new(kind, &String::from_utf8_lossy(&bs))
-                .with_context("response", format!("{parts:?}"));
+    if retryable {
+        err = err.set_temporary();
+    }
 
-            if retryable {
-                err = err.set_temporary();
-            }
+    if let Some(uri) = parts.extensions.get::<Uri>() {
+        err = err.with_context("uri", uri.to_string());
+    }
 
-            Ok(err)
-        }
+    Ok(err)
+}
+
+/// We cannot get the error type from the response header when the status code is 409.
+/// Because Dropbox API v2 will put error summary in the response body,
+/// we need to parse it to get the correct error type and then error kind.
+///
+/// See <https://www.dropbox.com/developers/documentation/http/documentation#error-handling>
+pub fn parse_dropbox_error_summary(summary: &str) -> Option<(ErrorKind, bool)> {
+    if summary.starts_with("path/not_found") || summary.starts_with("path_lookup/not_found") {
+        Some((ErrorKind::NotFound, false))
+    } else if summary.starts_with("path/conflict") {
+        Some((ErrorKind::AlreadyExists, false))
+    } else {
+        None
     }
 }
