--- core/src/services/dropbox/response.rs
@@ -1,86 +1,0 @@
-// Licensed to the Apache Software Foundation (ASF) under one
-// or more contributor license agreements.  See the NOTICE file
-// distributed with this work for additional information
-// regarding copyright ownership.  The ASF licenses this file
-// to you under the Apache License, Version 2.0 (the
-// "License"); you may not use this file except in compliance
-// with the License.  You may obtain a copy of the License at
-//
-//   http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing,
-// software distributed under the License is distributed on an
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-// KIND, either express or implied.  See the License for the
-// specific language governing permissions and limitations
-// under the License.
-
-use serde::Deserialize;
-
-#[derive(Default, Debug, Deserialize)]
-#[serde(default)]
-pub struct DropboxErrorResponse {
-    pub error_summary: String,
-    pub error: DropboxErrorDetail,
-}
-
-#[derive(Default, Debug, Deserialize)]
-#[serde(default)]
-pub struct DropboxErrorDetail {
-    #[serde(rename(deserialize = ".tag"))]
-    pub tag: String,
-}
-
-#[derive(Default, Debug, Deserialize)]
-#[serde(default)]
-pub struct DropboxMetadataResponse {
-    #[serde(rename(deserialize = ".tag"))]
-    pub tag: String,
-    pub client_modified: String,
-    pub content_hash: Option<String>,
-    pub file_lock_info: Option<DropboxMetadataFileLockInfo>,
-    pub has_explicit_shared_members: Option<bool>,
-    pub id: String,
-    pub is_downloadable: Option<bool>,
-    pub name: String,
-    pub path_display: String,
-    pub path_lower: String,
-    pub property_groups: Option<Vec<DropboxMetadataPropertyGroup>>,
-    pub rev: Option<String>,
-    pub server_modified: Option<String>,
-    pub sharing_info: Option<DropboxMetadataSharingInfo>,
-    pub size: Option<u64>,
-}
-
-#[derive(Default, Debug, Deserialize)]
-#[serde(default)]
-pub struct DropboxMetadataFileLockInfo {
-    pub created: Option<String>,
-    pub is_lockholder: bool,
-    pub lockholder_name: Option<String>,
-}
-
-#[derive(Default, Debug, Deserialize)]
-#[serde(default)]
-pub struct DropboxMetadataPropertyGroup {
-    pub fields: Vec<DropboxMetadataPropertyGroupField>,
-    pub template_id: String,
-}
-
-#[derive(Default, Debug, Deserialize)]
-#[serde(default)]
-pub struct DropboxMetadataPropertyGroupField {
-    pub name: String,
-    pub value: String,
-}
-
-#[derive(Default, Debug, Deserialize)]
-#[serde(default)]
-pub struct DropboxMetadataSharingInfo {
-    pub modified_by: Option<String>,
-    pub parent_shared_folder_id: Option<String>,
-    pub read_only: Option<bool>,
-    pub shared_folder_id: Option<String>,
-    pub traverse_only: Option<bool>,
-    pub no_access: Option<bool>,
-}
